Break import loop
