+++ dace/codegen/instrumentation/papi.py
@@ -5,7 +5,7 @@
 import dace
 from dace import dtypes, registry, symbolic
 from dace.codegen.instrumentation.provider import InstrumentationProvider
-from dace.codegen.targets.common import sym2cpp
+from dace.codegen.common import sym2cpp
 from dace.config import Config
 from dace.sdfg import nodes
 from dace.sdfg.nodes import EntryNode, MapEntry, MapExit, Tasklet
